Rename CompressionCallback to OpenVINOCallback
