+++ tests/pre_merge/utils/callbacks/openvino_callback/dummy_config.yml
